+++ packages/dialect-react-ui/components/Notifications/index.tsx
@@ -182,7 +182,10 @@ export default function Notifications({
       >
         <Router>
           <WalletStatesWrapper header={fallbackHeader}>
-            <ConnectionWrapper header={fallbackHeader}>
+            <ConnectionWrapper
+              header={fallbackHeader}
+              pollingInterval={props.pollingInterval}
+            >
               <GatedWrapper gatedView={gatedView}>
                 <ThreadEncyprionWrapper otherMemberPK={dappAddress}>
                   <InnerNotifications {...props} />