+++ packages/dialect-react-ui/components/common/providers/DialectUiManagementProvider.tsx
@@ -1,6 +1,5 @@
 import React, {
   createContext,
-  FunctionComponent,
   useCallback,
   useContext,
   useEffect,